requirements.txt
pygame==1.9.6
Pillow==6.2.0
psutil==5.6.3
